// apps/dashboard/api/src/modules/auth/error/AuthErrorCode.ts
enum AuthErrorCode {
  EMAIL_ALREADY_EXISTS = "EMAIL_ALREADY_EXISTS",
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  INVALID_REFRESH_TOKEN = "INVALID_REFRESH_TOKEN",
  EXPIRED_SESSION = "EXPIRED_SESSION",
  REVOKED_SESSION = "REVOKED_SESSION",
  MISSING_ACCESS_TOKEN = "MISSING_ACCESS_TOKEN",
  INVALID_ACCESS_TOKEN = "INVALID_ACCESS_TOKEN",
}

export default AuthErrorCode;